majorproject/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Kafka Client
confluent-kafka

# PostgreSQL Client
psycopg2-binary

# Configuration & Environment
python-dotenv
pyyaml

# Data Generation & Testing
faker

# CLI Tools
click

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.1

# Logging & Monitoring
colorlog

# Development Tools
pytest
pytest-asyncio
httpx
pyyaml

pyspark
redis>=4.6.0
colorama
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
prometheus-client>=0.20.0